GridWorld
GridWorld is a world where everybody is a number and lives in a square (unless of course they are dead).
The world is an RxC grid (rows and columns). Each entry on the grid is called a ​district ​and is referred to by its row ​r ​and column ​c​ where ​r∈{0..R-1}​ and ​c∈{0..C-1}​. 
People live in the world; ​each uniquely identified by an integer ID​ (like a social security number). Person IDs start at zero. Each living person resides in one (and only one) of the districts.